+++ include/expressionEngine/parser/functionParser.h
@@ -4,6 +4,7 @@
 #include "core/errorHandler.h"
 #include "expressionEngine/functionManager.h"
 
+enum reh_error_code_e ree_testLexerParser(void);
 enum reh_error_code_e ree_parseFunctionDefinition(char* definition, struct ree_function_t *function);
 
 #endif // FUNCTION_PARSER_H